+++ Source/Core/Request.swift
@@ -99,7 +99,8 @@ public struct Request {
     /// The body of the HTTP request.
     public var body: Data?
     
-
+    /// The Boolean to indicate if default cookies should be set for request.
+    public var shouldHandleCookies: Bool = true
     
     /// The key/value pairs that will be encoded as the query in the URL.
     public var queryParameters: [String : Any]?
@@ -197,6 +198,7 @@ extension Request: URLRequestEncodable {
         let urlRequest = NSMutableURLRequest(url: URL(string: url)!)
         urlRequest.httpMethod = method.rawValue
         urlRequest.cachePolicy = cachePolicy
+        urlRequest.httpShouldHandleCookies = shouldHandleCookies
         
         for (name, value) in headers {
             urlRequest.addValue(value, forHTTPHeaderField: name)