+++ Source/Core/ServiceTask.swift
@@ -37,6 +37,20 @@ import Foundation
 // MARK: - Request API
 
 extension ServiceTask {
+    
+    /**
+     Sets a boolean that indicates whether the receiver should use the default
+     cookie handling for the request.
+     
+     - parameter handle: true if the receiver should use the default cookie
+     handling for the request, false otherwise. The default is true.
+     - returns: Self instance to support chaining.
+     */
+    @discardableResult public func setShouldHandleCookies(_ handle: Bool) -> Self {
+        request.shouldHandleCookies = handle
+        return self
+    }
+    
     /// TODO: Needs docs
     @discardableResult public func setParameters(_ parameters: [String: Any], encoding: Request.ParameterEncoding? = nil) -> Self {
         request.parameters = parameters